learnflow-app/progress-api/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
openai-agents
